create the storyboards views by hand
