+++ Core Data Journal/Core Data Journal/Misc/Assets.xcassets/AppIcon.appiconset/Contents.json
