+++ Core Data Journal/Core Data Journal/Misc/Assets.xcassets/Contents.json
