+++ Core Data Journal/Core Data Journal/Misc/Base.lproj/LaunchScreen.storyboard
