+++ Core Data Journal/Core Data Journal/View Controllers/EntryDetailViewController.swift
@@ -1,0 +1,28 @@
+//
+//  EntryDetailViewController.swift
+//  Core Data Journal
+//
+//  Created by Jaspal on 2/18/19.
+//  Copyright © 2019 Jaspal Suri. All rights reserved.
+//
+
+import UIKit
+
+class EntryDetailViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        // Do any additional setup after loading the view.
+    }
+    
+
+    // MARK: - Navigation
+
+    // In a storyboard-based application, you will often want to do a little preparation before navigation
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        // Get the new view controller using segue.destination.
+        // Pass the selected object to the new view controller.
+    }
+
+}
