--- Core Data Journal/Core Data Journal/ViewController.swift
@@ -1,20 +1,0 @@
-//
-//  ViewController.swift
-//  Core Data Journal
-//
-//  Created by Jaspal on 2/18/19.
-//  Copyright © 2019 Jaspal Suri. All rights reserved.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view, typically from a nib.
-    }
-
-
-}
-
